+++ src/vue/modules/asset-explorer/wrappers/balance.js
@@ -1,6 +1,0 @@
-export class Balance {
-  constructor (record) {
-    this.available = record.state.available
-    this.assetCode = record.asset.id
-  }
-}