+++ src/vue/modules/assets/asset-explorer/_api.js
@@ -4,8 +4,8 @@ let _api = null
 
 /**
  * @param {Wallet} wallet - wallet to sign the requests
- * @param {Wallet} config - wallet to sign the requests
- * @param {Wallet} config.horizonURL - the url of the horizon server
+ * @param {Object} config
+ * @param {String} config.horizonURL - the url of the horizon server
  * (without version prefix)
  */
 export function initApi (wallet, config) {